Rename files to avoid collision
